+++ next-docs/components/facing/CardContibutor.tsx
@@ -37,7 +37,7 @@ const CardContributor: React.FC<Props> = ({
         variant && variant in variants ? variants[variant] : variants.default
       }
     >
-      <p className="list-item absolute bottom-12 pb-2 left-full whitespace-nowrap text-goten text-moon-24 transform origin-bottom-left -rotate-90 z-10 font-medium">
+      <p className="list-item absolute bottom-12 pb-2 ltr:left-full rtl:right-full whitespace-nowrap text-goten text-moon-24 transform origin-bottom-left -rotate-90 z-10 font-medium">
         {title}
       </p>
       <Image