+++ next-docs/public/examples/iconsTW/Other.tsx
@@ -3,17 +3,19 @@ import {
   Other3Dots,
   Other3DotsHorizontal,
   OtherFrame,
+  OtherFreebet,
   OtherId,
+  OtherLifebuoy,
   OtherLightning,
   OtherMoon,
   OtherOther,
   OtherPlug,
   OtherRewards,
+  OtherRocket,
   OtherSad,
   OtherSmile,
   OtherSun,
   OtherWater,
-  ClubhouseGames,
 } from '@heathmont/moon-icons-tw';
 import Icon from './Icon';
 
@@ -38,8 +40,14 @@ const Example = ({ wrapperProps, props }: Props) => (
     <Icon {...wrapperProps} name="OtherFrame">
       <OtherFrame {...props} />
     </Icon>
+    <Icon {...wrapperProps} name="OtherFreebet">
+      <OtherFreebet {...props} />
+    </Icon>
     <Icon {...wrapperProps} name="OtherId">
       <OtherId {...props} />
+    </Icon>
+    <Icon {...wrapperProps} name="OtherLifebuoy">
+      <OtherLifebuoy {...props} />
     </Icon>
     <Icon {...wrapperProps} name="OtherLightning">
       <OtherLightning {...props} />
@@ -56,6 +64,9 @@ const Example = ({ wrapperProps, props }: Props) => (
     <Icon {...wrapperProps} name="OtherRewards">
       <OtherRewards {...props} />
     </Icon>
+    <Icon {...wrapperProps} name="OtherRocket">
+      <OtherRocket {...props} />
+    </Icon>
     <Icon {...wrapperProps} name="OtherSad">
       <OtherSad {...props} />
     </Icon>
@@ -67,9 +78,6 @@ const Example = ({ wrapperProps, props }: Props) => (
     </Icon>
     <Icon {...wrapperProps} name="OtherWater">
       <OtherWater {...props} />
-    </Icon>
-    <Icon {...wrapperProps} name="ClubhouseGames">
-      <ClubhouseGames {...props} />
     </Icon>
   </>
 );