+++ next-docs/public/examples/loaderTW/Default.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Loader } from '@heathmont/moon-core-tw';
 
 const Example = () => <Loader />;