+++ packages/select/src/private/SelectGroupEnd.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { TextInput } from "@heathmont/moon-core";
+import { TextInput } from '@heathmont/moon-core';
 import styled from 'styled-components';
 import { Select } from '..';
-import {CombinedPropType} from "./types/CombinedInputProps";
-import {Size, TextInputGroupSize} from "./types/CombinedInputSize";
+import { CombinedPropType } from './types/CombinedInputProps';
+import { Size, TextInputGroupSize } from './types/CombinedInputSize';
 import { OrientationType, Orientation } from './types/OrientationType';
 
 const Wrapper = styled.div({
@@ -22,12 +22,12 @@ const SelectGroupEnd: React.FC<SelectGroupEndProps> = ({
   inputProps,
   orientation,
   isRtl,
-  size
+  size,
 }) => {
-  const selectSize = size === Size.medium ? Size.medium : Size.large;
-  const inputSize = size === Size.medium ? Size.large : Size.xlarge;
+  const inputSize = size === Size.medium ? Size.large : size;
+  const selectSize = size === Size.xlarge ? 'xLarge' : size;
 
-  if (inputProps && inputProps.type === "select") {
+  if (inputProps && inputProps.type === 'select') {
     return (
       <Wrapper>
         <Select
@@ -41,13 +41,13 @@ const SelectGroupEnd: React.FC<SelectGroupEndProps> = ({
           size={selectSize}
         />
       </Wrapper>
-    )
+    );
   }
 
   return (
     <Wrapper>
       <TextInput
-        {...(inputProps ?? { inputSize: 'medium', type: 'text' })}
+        {...(inputProps ?? { inputSize: 'xlarge', type: 'text' })}
         dir={isRtl ? 'rtl' : 'ltr'}
         isSharpRightSide={orientation === Orientation.horizontal && isRtl}
         isSharpLeftSide={orientation === Orientation.horizontal && !isRtl}