+++ workspaces/core/src/index.ts
@@ -22,8 +22,6 @@ export { default as setSpacing } from './label/utils/setSpacing';
 export * from './label/utils/setSpacing';
 export { default as Loader } from './loader/Loader';
 export * from './loader/Loader';
-export { default as Pagination } from './pagination/Pagination';
-export * from './pagination/Pagination';
 export { default as RadioButton } from './radioButton/RadioButton';
 export * from './radioButton/RadioButton';
 export { default as TextInput } from './textInput/TextInput';