+++ workspaces/core/src/radioButton/private/types/Props.ts
@@ -1,0 +1,11 @@
+type Props = React.InputHTMLAttributes<HTMLInputElement> & {
+  disabled?: boolean;
+  ariaLabel?: string;
+  label?: JSX.Element | string;
+  id?: string;
+  checked?: boolean;
+  onClick?: (event?: React.MouseEvent<HTMLInputElement>) => void;
+  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+};
+
+export default Props;