+++ workspaces/icons/src/icons/MediaSoundwave.tsx
@@ -10,9 +10,8 @@ const MediaSoundwave = (props: React.SVGProps<SVGSVGElement>) => (
     {...props}
   >
     <path
-      d="M16 25.5v-19m-4.75 15.133V10.367M6.5 18.54v-5.078m14.25 8.172V10.367m4.75 8.172v-5.078"
-      stroke="currentColor"
-      strokeLinecap="round"
+      d="M15.5 25.5a.5.5 0 101 0h-1zm1-19a.5.5 0 10-1 0h1zm-5.75 15.133a.5.5 0 001 0h-1zm1-11.266a.5.5 0 10-1 0h1zM6 18.54a.5.5 0 101 0H6zm1-5.078a.5.5 0 00-1 0h1zm13.25 8.172a.5.5 0 001 0h-1zm1-11.266a.5.5 0 10-1 0h1zM25 18.54a.5.5 0 101 0h-1zm1-5.078a.5.5 0 00-1 0h1zM16.5 25.5v-19h-1v19h1zm-4.75-3.867V10.367h-1v11.266h1zM7 18.539v-5.078H6v5.078h1zm14.25 3.094V10.367h-1v11.266h1zM26 18.539v-5.078h-1v5.078h1z"
+      fill="currentColor"
     />
   </svg>
 );