+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 9.2.1
+
+### Patch Changes
+
+- fixes
+
+## 9.2.0
+
+### Minor Changes
+
+- Release icons & pnpm7
+
 ## 9.1.8
 
 ### Patch Changes