+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 9.2.1
+
+### Patch Changes
+
+- fixes
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.2.1
+
+## 9.2.0
+
+### Minor Changes
+
+- Release icons & pnpm7
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.2.0
+
 ## 9.1.8
 
 ### Patch Changes